packages/stepproof-cc-adapter/pyproject.toml
[project]
name = "stepproof-cc-adapter"
version = "0.0.1"
description = "StepProof Claude Code adapter — uv PreToolUse hook, classification, subagents, slash commands"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "BUSL-1.1" }
authors = [{ name = "Eidos AGI", email = "daniel@eidosagi.com" }]

dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stepproof_cc_adapter"]
